IO 整体重构

+++ Test/TestMD5/Src/TestMD5.cpp
@@ -6,7 +6,7 @@ int main()
 {
     string src = "HelloWorldHelloWorldHelloWorldHelloWorldHelloWorldHelloWorld";
 
-    string s = getMD5((const unsigned char*)src.c_str(), src.length());
+    string s = getMD5((const unsigned char*)src.c_str(), (int)src.length());
     cout << s;
     return 0;
 }
